Massive refactor
